requirements.txt
# Core application dependencies
flask
requests
onnxruntime
scikit-learn
skl2onnx
numpy

# Testing dependencies
pytest>=6.0.0
pytest-mock
pytest-cov
pytest-timeout
playwright>=1.20.0
pytest-playwright

# Type checking dependencies
types-requests

# Phase 2: LangChain and gRPC dependencies
langchain>=0.1.0
langchain-community>=0.0.10
grpcio>=1.60.0
grpcio-tools>=1.60.0
joblib

# Phase 3: Advanced caching dependencies
Flask-Caching>=2.0.0
redis>=4.0.0
fastembed>=0.2.0

# Phase 4: Production model lifecycle & monitoring dependencies
evidently>=0.4.0
mlflow>=2.9.0
pandas>=1.5.0

# Phase 5: Dedicated model serving infrastructure dependencies
tensorflow>=2.10.0
tf2onnx>=1.14.0
tensorflow-serving-api>=2.10.0
tritonclient[http]>=2.37.0